src/requirements.txt
Django==1.10.5
Pygments==2.1.3
Werkzeug==0.11.15
celery==4.0.2
coverage==4.3.4
django-environ==0.4.1
django-extensions==1.7.5
django-nose==1.4.4
django-redis-cache==1.7.1
gunicorn==19.6.0
model-mommy==1.3.1
opbeat==3.5.2
psycopg2==2.6.2
redis==2.10.5
sendgrid-django==4.0.4
sqlparse==0.2.2
